# internal/apidocsgen/openapi.template.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: MediaMTX API
  description: API of MediaMTX, a server and proxy that supports various protocols.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9997
security: []
...
paths:
  /v3/info:
    get:
      operationId: info
      tags: [General]
      summary: returns informations about the instance.
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Info"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/auth/jwks/refresh:
    post:
      operationId: authJwksRefresh
      tags: [Authentication]
      summary: Manually refreshes the JWT JWKS.
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/config/global/get:
    get:
      operationId: configGlobalGet
      tags: [Configuration]
      summary: returns the global configuration.
      description: ""
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalConf"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/config/global/patch:
    patch:
      operationId: configGlobalSet
      tags: [Configuration]
      summary: patches the global configuration.
      description: all fields are optional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalConf"
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/config/path-defaults/get:
    get:
      operationId: configPathDefaultsGet
      tags: [Configuration]
      summary: returns the default path configuration.
      description: ""
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathConf"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/config/path-defaults/patch:
    patch:
      operationId: configPathDefaultsPatch
      tags: [Configuration]
      summary: patches the default path configuration.
      description: all fields are optional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathConf"
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/config/paths/list:
    get:
      operationId: configPathsList
      tags: [Configuration]
      summary: returns all path configurations.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathConfList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v3/hls/muxers/list:
    get:
      operationId: hlsMuxersList
      tags: [HLS]
      summary: returns all HLS muxers.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HLSMuxerList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v3/hls/sessions/list:
    get:
      operationId: hlssessionsList
      tags: [HLS]
      summary: returns all HLS sessions.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HLSSessionList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/hls/sessions/get/{id}:
    get:
      operationId: hlssessionsGet
      tags: [HLS]
      summary: returns a HLS session.
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the session.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HLSSession"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: session not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/hls/sessions/kick/{id}:
    post:
      operationId: hlssessionsKick
      tags: [HLS]
      summary: kicks out a HLS session from the server.
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the session.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: session not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/paths/list:
    get:
      operationId: pathsList
      tags: [Paths]
      summary: returns all paths.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v3/paths/forward-dests/list:
    get:
      operationId: pathsForwardDestsList
      tags: [Paths]
      summary: returns all forward destinations of a path.
      description: ""
      parameters:
        - name: path
          in: query
          required: true
          description: name of the path.
          schema:
            type: string
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForwardDestList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: path not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/paths/forward-dests/get:
    get:
      operationId: pathsForwardDestsGet
      tags: [Paths]
      summary: returns a forward destination.
      description: ""
      parameters:
        - name: id
          in: query
          required: true
          description: ID of the forward destination.
          schema:
            type: string
            format: uuid
        - name: path
          in: query
          required: true
          description: name of the path.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForwardDest"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: path or forward destination not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v3/rtsp/conns/list:
    get:
      operationId: rtspConnsList
      tags: [RTSP]
      summary: returns all RTSP connections.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPConnList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsp/conns/get/{id}:
    get:
      operationId: rtspConnsGet
      tags: [RTSP]
      summary: returns a RTSP connection.
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the connection.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPConn"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsp/sessions/list:
    get:
      operationId: rtspSessionsList
      tags: [RTSP]
      summary: returns all RTSP sessions.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPSessionList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsp/sessions/get/{id}:
    get:
      operationId: rtspSessionsGet
      tags: [RTSP]
      summary: returns a RTSP session.
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the connection.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPSession"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: session not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsp/sessions/kick/{id}:
    post:
      operationId: rtspSessionsKick
      tags: [RTSP]
      summary: kicks out a RTSP session from the server.
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the session.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: session not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsps/conns/list:
    get:
      operationId: rtspsConnsList
      tags: [RTSP]
      summary: returns all RTSPS connections.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPConnList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsps/conns/get/{id}:
    get:
      operationId: rtspsConnsGet
      tags: [RTSP]
      summary: returns a RTSPS connection.
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the connection.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPConn"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsps/sessions/list:
    get:
      operationId: rtspsSessionsList
      tags: [RTSP]
      summary: returns all RTSPS sessions.
      description: ""
      parameters:
        - name: page
          in: query
          description: page number.
          schema:
            type: integer
            default: 0
        - name: itemsPerPage
          in: query
          description: items per page.
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPSessionList"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v3/rtsps/sessions/get/{id}:
    get:
      operationId: rtspsSessionsGet
      tags: [RTSP]
      summary: returns a RTSPS session.
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the connection.
          schema:
            type: string
      responses:
        "200":
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RTSPSession"
        "400":
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: session not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...